* break out cli tests into files clearly delimited by backend if appropriate * lots of linting
