+++ tests/cli/__init__.py
@@ -1,0 +1,2 @@
+# Since our cli produces unicode output, but we want tests in python2 as well
+from __future__ import unicode_literals
